% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{business_search}
\alias{business_search}
\title{search businesse on yelp}
\usage{
business_search(api_key, location = NULL, latitude = NULL,
  longitude = NULL, term = NULL, ...)
}
\arguments{
\item{api_key}{string}

\item{location}{string. Required if either latitude or longitude is not provided. Specifies the combination of "address, neighborhood, city, state or zip, optional country" to be used when searching for businesses}

\item{latitude}{decimal. Required if location is not provided. Latitude of the location you want to search nearby}

\item{longitude}{decimal. Required if location is not provided. Longitude of the location you want to search nearby}

\item{term}{string. Optional search term (e.g. "food", "restaurants"). If term isn’t included we search everything. The term keyword also accepts business names such as "Starbucks"}

\item{...}{optional paramters \href{https://www.yelp.com/developers/documentation/v3/business_search}{More}}
}
\value{
List with the following elements:
\itemize{
\item total: Total number of business Yelp finds based on the search criteria. Sometimes, the value may exceed 1000. In such case, you still can only get up to 1000 businesses. total may be limited to 40 for non-default sorts such as "distance" and "review_count"
\item businesses:  List of business Yelp finds based on the search criteria
\item region: Suggested area in a map to display results in
}
}
\description{
search businesse on yelp
}
\examples{
# search businesses with keyword 'food' in New York
key <- "######"
business_search(key,
                location = 'New York',
                term = "chinese")

}
