% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{business_search_phone}
\alias{business_search_phone}
\title{Return a list of businesses on yelp based on the provided phone number}
\usage{
business_search_phone(api_key, phone_number)
}
\arguments{
\item{api_key}{string}

\item{phone_number}{Phone number of the business you want to search for. It must start with + and include the country code, like +14159083801}
}
\value{
List with the following elements:
\itemize{
\item total: Total number of business Yelp finds based on the search criteria. Sometimes, the value may exceed 1000. In such case, you still can only get up to 1000 businesses. total may be limited to 40 for non-default sorts such as "distance" and "review_count"
\item businesses:  List of business Yelp finds based on the search criteria
}
}
\description{
Return a list of businesses on yelp based on the provided phone number
}
\examples{
key <- "######"
business_search_phone(api_key = key, phone_number = '+14159083801')

}
