% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_events_endpoints.R
\name{event_search}
\alias{event_search}
\title{return events based on the provided search criteria}
\usage{
event_search(api_key, ...)
}
\arguments{
\item{api_key}{string}

\item{...}{optional parameters \href{https://www.yelp.com/developers/documentation/v3/event_search}{More}}
}
\value{
List with the following elements:
\itemize{
\item total: Total number of events returned based on search criteria
\item events: List of events found matching search criteria
}
}
\description{
return events based on the provided search criteria
}
\examples{
key <- "######"
## retrieve 10 events
event_search(key, limit = 10)

}
