% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hours.R
\name{get_hours}
\alias{get_hours}
\title{Print a table of hours for a business}
\usage{
get_hours(tbl, business_name = NULL)
}
\arguments{
\item{tbl}{The results of \code{yelp_search} stored as a dataframe.}

\item{business_name}{Optional string. The name of a specific business to get hours for. Defaults to \code{NULL}.}
}
\value{
A dataframe of hours of operation.
}
\description{
This is a helper function to be used with \code{yelp_search} to parse the hours for a specific business or list of businesses.
}
\examples{
get_hours(tbl)
get_hours(tbl, "Wegmans")
}
