% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-client.R
\name{get_secret}
\alias{get_secret}
\title{Get and store your Yelp Fusion client secret}
\usage{
get_secret(client_secret)
}
\arguments{
\item{client_secret}{Your Yelp API client secret generated through \href{https://www.yelp.com/developers/documentation/v3/authentication}{Yelp Fusion}.}
}
\value{
Returns and stores the value of your client secret.
}
\description{
This function confirms that the Yelp Fusion client secret is valid and stored correctly in your keyring.
}
