% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelp_search.R
\name{get_yelp_search_data}
\alias{get_yelp_search_data}
\title{Search yelp business data}
\usage{
get_yelp_search_data(term = NULL, location = NULL, latitude = NULL,
  longitude = NULL, radius = NULL, categories = NULL,
  locale = NULL, limit = 50, price = NULL, attributes = NULL,
  client_secret = NULL)
}
\arguments{
\item{term}{Search term}

\item{location}{Search location. City with two letter state abbreviation. EX. \code{"Charlottesville, VA"}}

\item{latitude}{The geographic latitude measurement}

\item{longitude}{The geographic measurement}

\item{radius}{Size of geographic area to pull results from}

\item{categories}{Result categories. EX. \code{"Restaurants", "Health & Fitness", "Automobile"}}

\item{locale}{Language and country codes, from \url{https://www.yelp.com/developers/documentation/v3/supported_locales}}

\item{limit}{Maximum number of results}

\item{price}{Price ranging from $ to $$$}

\item{client_secret}{Your personal client secret}
}
\description{
With this function you can search all businesses in the yelp directory based on geographic location.
}
\details{
The output is a large table containing relevant businesses.
}
