% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_header}
\alias{prepare_header}
\title{prepare header by appending "Bearer" to the api_key}
\usage{
prepare_header(api_key)
}
\arguments{
\item{api_key}{string}
}
\description{
prepare header by appending "Bearer" to the api_key
}
