% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tibble}
\alias{print.tibble}
\title{Print a table of the business listings}
\usage{
\method{print}{tibble}(x, ..., view = interactive())
}
\arguments{
\item{x}{Yelp business search results.}

\item{...}{Optional arguments to print.}

\item{view}{Interactive}
}
\description{
Print a table of the business listings
}
