% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{search_autocomplete}
\alias{search_autocomplete}
\title{return autocomplete suggestions for search keywords, businesses and categories, based on the input text}
\usage{
search_autocomplete(api_key, input_text)
}
\description{
return autocomplete suggestions for search keywords, businesses and categories, based on the input text
}
\examples{
key <- "######"
search_autocomplete(key, "star")

}
