% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{search_business}
\alias{search_business}
\title{search businesse on yelp}
\usage{
search_business(api_key, parameters)
}
\arguments{
\item{api_key}{string}

\item{parameters}{list corresponding to API paremeters  \url{https://www.yelp.com/developers/documentation/v3/business_search}}
}
\value{
list with the following elements:
\itemize{
\item total: Total number of business Yelp finds based on the search criteria. Sometimes, the value may exceed 1000. In such case, you still can only get up to 1000 businesses. total may be limited to 40 for non-default sorts such as "distance" and "review_count"
\item businesses:  List of business Yelp finds based on the search criteria
\item region: Suggested area in a map to display results in
}
}
\description{
search businesse on yelp
}
\details{
\itemize{
\item location(string): Required if either latitude or longitude is not provided. Specifies the combination of "address, neighborhood, city, state or zip, optional country" to be used when searching for businesses
\item latitude(decimal): Required if location is not provided. Latitude of the location you want to search nearby
\item longitude(decimal): Required if location is not provided. Longitude of the location you want to search nearby.
\item optional parameters ..
}
}
\examples{
# search businesses with keyword 'food' in New York
key <- "######"
search_business(api_key = key, parameters = list(term ='food',
                                                 location = 'New York'))

}
