% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{search_business_review}
\alias{search_business_review}
\title{return up to three review excerpts for a given business ordered by Yelp's default sort order}
\usage{
search_business_review(api_key, business_id)
}
\description{
return up to three review excerpts for a given business ordered by Yelp's default sort order
}
\examples{
key <- "######"
search_business_review(key, 'gary-danko-san-francisco')

}
