% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelp_business.R
\name{yelp_business}
\alias{yelp_business}
\title{Connect to Yelp and get data}
\usage{
yelp_business(id, client_secret = yelp_key("yelp"))
}
\arguments{
\item{id}{The business id provided by Yelp}

\item{client_secret}{Your Yelp API client secret}
}
\description{
With this function you can connect to Yelp using your API client secret or keyring.
}
\details{
The output is a large table containing relevant businesses.
}
