% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelp_businesses.R
\name{yelp_businesses}
\alias{yelp_businesses}
\title{Connect to Yelp and get a table of businesses}
\usage{
yelp_businesses(ids, client_secret = yelp_key("yelp"))
}
\arguments{
\item{ids}{The business ids provided by Yelp}

\item{client_secret}{Your Yelp API client secret}
}
\description{
This function connects to the Yelp API and gets data for a relevant business based on the business id.
}
\details{
The output is
}
