% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelp_search.R
\name{yelp_key}
\alias{yelp_key}
\title{Assign "yelp_key" class to your Yelp API key}
\usage{
yelp_key(key)
}
\arguments{
\item{key}{yelp key}
}
\description{
This is an internal function for assigning the custom class "yelp_key" to the key value.
}
