% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{business_search_review}
\alias{business_search_review}
\title{return up to three review excerpts for a given business ordered by Yelp's default sort order}
\usage{
business_search_review(api_key, business_id)
}
\arguments{
\item{api_key}{string}

\item{business_id}{business id}
}
\description{
return up to three review excerpts for a given business ordered by Yelp's default sort order
}
\examples{
key <- "######"
business_search_review(key, 'gary-danko-san-francisco')

}
