% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_business_endpoints.R
\name{business_search_transaction}
\alias{business_search_transaction}
\title{return a list of businesses which support food delivery transactions.}
\usage{
business_search_transaction(api_key, transaction_type = "delivery",
  parameters)
}
\description{
return a list of businesses which support food delivery transactions.
}
\examples{
key <- "######"
business_search_transaction(key,
                           transaction_type = "delivery",
                           list(longitude = "-122.399305736113", latitude = "37.787789124691"))

}
