% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_events_endpoints.R
\name{event_lookup_id}
\alias{event_lookup_id}
\title{return the detailed information of a Yelp event}
\usage{
event_lookup_id(api_key, event_id)
}
\arguments{
\item{api_key}{string}

\item{event_id}{event id}
}
\description{
return the detailed information of a Yelp event
}
\details{
This endpoint requires you be in yelp developer beta
}
\examples{
key <- "######"
event_lookup_id(key, "oakland-saucy-oakland-restaurant-pop-up")

}
