% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_events_endpoints.R
\name{event_search}
\alias{event_search}
\title{return events based on the provided search criteria.}
\usage{
event_search(api_key, parameters = list())
}
\description{
return events based on the provided search criteria.
}
\examples{
key <- "######"
## retrieve 10 events
event_search(key, list(limit = 10))

}
