% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpr_events_endpoints.R
\name{event_search_featured}
\alias{event_search_featured}
\title{return the featured event for a given location. Featured events are chosen by Yelp's community managers.}
\usage{
event_search_featured(api_key, parameters)
}
\description{
return the featured event for a given location. Featured events are chosen by Yelp's community managers.
}
\examples{
key <- "######"
event_search_featured(api_key = key, parameters = list(location = 'New York'))

}
