% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hours.R
\name{get_hours}
\alias{get_hours}
\title{Print a table of hours for a business}
\usage{
get_hours(tbl, client_secret = yelp_key("yelp"))
}
\arguments{
\item{tbl}{}

\item{client_secret}{}
}
\description{
Print a table of hours for a business
}
