% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hours.R
\name{get_hours}
\alias{get_hours}
\title{Print a table of hours for a business}
\usage{
get_hours(tbl, business_name = NULL)
}
\arguments{
\item{tbl}{the results of yelp_search}
}
\description{
This is a helper function
}
