% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_reviews}
\alias{get_reviews}
\title{Main function at user level}
\usage{
get_reviews(base_url, page_lim = NULL, company = NULL,
  verbose = TRUE)
}
\arguments{
\item{base_url}{url provided by user}

\item{page_lim}{number of pages to parse}

\item{company}{company name}

\item{verbose}{whether to print current url}
}
\description{
Main function at user level
}
\examples{
get_reviews("https://www.yelp.dk/biz/mad-og-kaffe-k\%C3\%B8benhavn-2","Mad&Kaffe")

}
