% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reviews.R
\name{get_reviews}
\alias{get_reviews}
\title{Print a table of three reviews for each business}
\usage{
get_reviews(tbl, business_name = NULL)
}
\arguments{
\item{tbl}{}

\item{business_name}{}
}
\description{
Print a table of three reviews for each business
}
