% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-client.R
\name{get_secret}
\alias{get_secret}
\title{Client Secret}
\usage{
get_secret(client_secret)
}
\arguments{
\item{client_secret}{Your Yelp API client secret}
}
\value{
client secret
}
\description{
This function confirms that the Yelp client secret is valid.
}
