% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelp_search.R
\name{get_yelp_search_data}
\alias{get_yelp_search_data}
\title{Search yelp business data}
\usage{
get_yelp_search_data(term = NULL, location = NULL, latitude = NULL,
  longitude = NULL, radius = NULL, categories = NULL,
  locale = NULL, limit = 50, price = NULL, attributes = NULL,
  client_secret = yelp_key("yelp"))
}
\arguments{
\item{term}{Search term as a string, for example \code{"coffee"} or
\code{"restaurants"}. The term may also be business names, such as
\code{"Starbucks"}.}

\item{location}{Search location as a string. City with two letter state
abbreviation. EX. \code{"Charlottesville, VA"}. Required if either latitude or
longitude is not provided.}

\item{latitude, longitude}{The geographic latitude and longitude measurements as a decimal. EX. \code{38.1}.
Required if \code{location} is not provided. If provided \code{location}, will be ignored.}

\item{radius}{Optional integer. A suggested search radius in meters. The max
value is 40,000 meters (about 25 miles).}

\item{categories}{Optional string. Categories to filter the search results
with, such as \code{"automotive"} or \code{"retail"}.
\href{https://www.yelp.com/developers/documentation/v3/all_category_list}{See
list of supported categories}.}

\item{locale}{Optional string. Specify the language and country codes.
Defaults to \code{en_US}.
\href{https://www.yelp.com/developers/documentation/v3/all_category_list}{See
list of supported locales}.}

\item{limit}{Optional integer. Number of business results to return. Defaults
to \code{20} and maximum is \code{50}.}

\item{price}{Optional string. Price categories to filter the search results
with, ranging from \code{"$"} to \code{"$$$$"}}

\item{attributes}{Optional string. Additional filters to the search results
such as \code{"wheelchair_accessible"}, or \code{"reservations"}.}

\item{client_secret}{Required. Your personal Yelp API key stored in your
keychain and retrieved by \code{get_key(service = "yelp")}.}
}
\description{
With this function you can search all businesses in the Yelp directory based
on geographic location.
}
\details{
The output is a large table containing relevant businesses.
}
