% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.business_tbl}
\alias{print.business_tbl}
\title{Print a leaflet map of results}
\usage{
\method{print}{business_tbl}(x, ..., view = interactive())
}
\arguments{
\item{x}{Yelp business table created from search query}
}
\description{
Print a leaflet map of results
}
